refactor: extract shared useActionHandlerContext hook
Consolidate common hook dependencies (socket, location, navigate,
translation, stores) into a shared hook used by both
ActionHandlerDecorator and BulkActionHandlerDecorator to reduce
code duplication.
